+++ src/components/Table.svelte
@@ -11,16 +11,23 @@
    */
 
   export let departureList: IDeparture[]
+  export let showETD: string
 </script>
 
 <table>
   <thead>
     <tr>
-      <th>Avgang</th>
-      <th>Rute</th>
+      {#if showETD === 'true'}
+        <th class="shrink center">Avgang</th>
+      {/if}
+      <th class="shrink">Rute</th>
       <th></th>
       <th>Destinasjon</th>
-      <th>Går om</th>
+      {#if showETD === 'true'}
+        <th>Går om</th>
+      {:else}
+        <th class="center">Avgang</th>
+      {/if}
     </tr>
   </thead>
   <tbody>
@@ -28,9 +35,11 @@
       <!-- Display only if time to departure is positive -->
       {#if minutesToDeparture(departure.departureTime) > 0}
         <tr>
-          <td>{ departure.departureTime.toLocaleTimeString('nb-NO', { hour: '2-digit', minute: '2-digit' }) }</td>
-          <td>{ departure.line }</td>
-          <td>
+          {#if showETD === 'true'}
+            <td class="shrink center">{ departure.departureTime.toLocaleTimeString('nb-NO', { hour: '2-digit', minute: '2-digit' }) }</td>
+          {/if}
+          <td class="shrink">{ departure.line }</td>
+          <td class="center">
             {#if departure.transport === 'air'}
               Fly
             {:else if departure.transport === 'bus'}
@@ -83,7 +92,11 @@
             {/if}
           </td>
           <td>{ departure.destination }</td>
-          <td><strong>{ countdownMessage(departure.departureTime) }</strong></td>
+          {#if showETD === 'true'}
+            <td>{ countdownMessage(departure.departureTime) }</td>
+          {:else}
+            <td class="center">{ departure.departureTime.toLocaleTimeString('nb-NO', { hour: '2-digit', minute: '2-digit' }) }</td>
+          {/if}
         </tr>
       {/if}
     {/each}
@@ -107,17 +120,12 @@
   th {
     padding: .6rem 1rem;
   }
-  td:first-child,
-  th:first-child,
-  th:nth-child(2),
-  td:nth-child(2) {
+  td.center,
+  th.center {
     text-align: center;
-    white-space: nowrap;
-    width: 0.1%;
   }
-  td:last-child,
-  th:last-child,
-  th:nth-child(3) {
+  td.shrink,
+  th.shrink {
     white-space: nowrap;
     width: 0.1%;
   }