+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Implement features requested by the client/custormer.
 
 ### Changes
 
+- Implemented a new table display mode.
 - Implemented a large departure time countdown at the bottom of the screen.
 
 ## v1.0.0-alpha.3